Round number of iters per epoch

+++ docs/examples/pytorch/main.py
@@ -248,7 +248,7 @@ def train(train_loader, model, criterion, optimizer, epoch):
     for i, data in enumerate(train_loader):
         input = data[0][0][0]
         target = data[0][1][0].cuda().long()
-        train_loader_len = train_loader._size / args.batch_size
+        train_loader_len = int(train_loader._size / args.batch_size)
 
         adjust_learning_rate(optimizer, epoch, i, train_loader_len)
 
@@ -322,7 +322,7 @@ def validate(val_loader, model, criterion):
     for i, data in enumerate(val_loader):
         input = data[0][0][0]
         target = data[0][1][0].cuda().long()
-        val_loader_len = val_loader._size / args.batch_size
+        val_loader_len = int(val_loader._size / args.batch_size)
 
         target = target.cuda(async=True)
         input_var = Variable(input)
